+++ frontend/src/types/messstelle/auswertung/MessstelleAuswertungOptionsDTO.ts
